feat: Add MangaBaka Tracker
Adds the MangaBaka tracking service, ported from Mihon PR #3047. It adds tracking preferences, tracking integration in track menus and search functionality, S256 PKCE code generation logic and the new MangaBaka UI sections in settings. References to Mihon have been updated to Neko.

+++ app/src/main/java/eu/kanade/tachiyomi/data/track/mangabaka/dto/MangaBakaListEntry.kt
@@ -15,7 +15,7 @@ data class MangaBakaListEntry(
     @SerialName("progress_chapter") val progressChapter: Double?,
     val rating: Long?,
 ) {
-    fun getStatus(): Int =
+    fun getStatus(): Long =
         when (state) {
             "considering" -> MangaBaka.CONSIDERING
             "completed" -> MangaBaka.COMPLETED
